// src/types/enums/KeyOptions.ts
export enum KeyOptions {
    AC = 'AC',
    INVERT_SIGNAL = '+/-',
    PERCENTAGE = '%',
    DIVISION = '/',
    SEVEN = '7',
    EIGHT = '8',
    NINE = '9',
    MULTIPLICATION = 'x',
    FOUR = '4',
    FIVE = '5',
    SIX = '6',
    MINUS = '-',
    ONE = '1',
    TWO = '2',
    THREE = '3',
    PLUS = '+',
    ZERO = '0',
    COMMA = '.',
    EQUAL = '=',
}